feat: 장소 단일 조회 & 장소 목록 조회 엔드포인트 분리
